Remove some dependencies in `chia.server`.
