+++ chia/_tests/cmds/test_farm_cmd.py
@@ -8,10 +8,12 @@ from _pytest.capture import CaptureFixture
 from chia._tests.util.time_out_assert import time_out_assert
 from chia.cmds.farm_funcs import summary
 from chia.farmer.farmer import Farmer
+from chia.farmer.farmer_service import FarmerService
 from chia.harvester.harvester import Harvester
-from chia.server.aliases import FarmerService, HarvesterService, WalletService
+from chia.harvester.harvester_service import HarvesterService
 from chia.simulator.block_tools import BlockTools
 from chia.simulator.start_simulator import SimulatorFullNodeService
+from chia.wallet.wallet_service import WalletService
 
 
 @pytest.mark.anyio
