+++ chia/harvester/start_harvester.py
@@ -13,8 +13,8 @@ from chia.consensus.default_constants import DEFAULT_CONSTANTS, update_testnet_o
 from chia.harvester.harvester import Harvester
 from chia.harvester.harvester_api import HarvesterAPI
 from chia.harvester.harvester_rpc_api import HarvesterRpcApi
+from chia.harvester.harvester_service import HarvesterService
 from chia.protocols.outbound_message import NodeType
-from chia.server.aliases import HarvesterService
 from chia.server.resolve_peer_info import get_unresolved_peer_infos
 from chia.server.signal_handlers import SignalHandlers
 from chia.server.start_service import RpcInfo, Service, async_run
