+++ chia/introducer/start_introducer.py
@@ -5,19 +5,13 @@ import pathlib
 import sys
 from typing import Any, Optional
 
-from chia_rs import ConsensusConstants
-
 from chia.apis import ApiProtocolRegistry
-from chia.consensus.constants import replace_str_to_bytes
-from chia.consensus.default_constants import DEFAULT_CONSTANTS, update_testnet_overrides
+from chia.introducer.introducer import Introducer
+from chia.introducer.introducer_api import IntroducerAPI
+from chia.introducer.introducer_service import IntroducerService
 from chia.protocols.outbound_message import NodeType
-from chia.server.aliases import TimelordService
-from chia.server.resolve_peer_info import get_unresolved_peer_infos
 from chia.server.signal_handlers import SignalHandlers
-from chia.server.start_service import RpcInfo, Service, async_run
-from chia.timelord.timelord import Timelord
-from chia.timelord.timelord_api import TimelordAPI
-from chia.timelord.timelord_rpc_api import TimelordRpcApi
+from chia.server.start_service import Service, async_run
 from chia.util.chia_logging import initialize_service_logging
 from chia.util.config import load_config, load_config_cli
 from chia.util.default_root import resolve_root_path
@@ -26,40 +20,45 @@ from chia.util.task_timing import maybe_manage_task_instrumentation
 # See: https://bugs.python.org/issue29288
 "".encode("idna")
 
-SERVICE_NAME = "timelord"
+SERVICE_NAME = "introducer"
 
 
-def create_timelord_service(
+def create_introducer_service(
     root_path: pathlib.Path,
     config: dict[str, Any],
-    consensus_constants: ConsensusConstants,
+    advertised_port: Optional[int] = None,
     connect_to_daemon: bool = True,
-) -> TimelordService:
+) -> IntroducerService:
     service_config = config[SERVICE_NAME]
 
     network_id = service_config["selected_network"]
-    overrides = service_config["network_overrides"]["constants"][network_id]
-    update_testnet_overrides(network_id, overrides)
-    updated_constants = replace_str_to_bytes(consensus_constants, **overrides)
 
-    node = Timelord(root_path, service_config, updated_constants)
-    peer_api = TimelordAPI(node)
+    if advertised_port is None:
+        advertised_port = service_config["port"]
 
-    rpc_info: Optional[RpcInfo[TimelordRpcApi]] = None
-    if service_config.get("start_rpc_server", True):
-        rpc_info = (TimelordRpcApi, service_config.get("rpc_port", 8557))
+    try:
+        default_port = service_config["network_overrides"]["config"][network_id]["default_full_node_port"]
+    except KeyError:
+        raise Exception(f"Specify default_full_node_port for network {network_id}")
+
+    dns_servers = service_config.get("dns_servers", [])
+    if dns_servers == [] and network_id == "mainnet":
+        dns_servers.append("dns-introducer.chia.net")
+
+    node = Introducer(
+        service_config["max_peers_to_send"], service_config["recent_peer_threshold"], default_port, dns_servers
+    )
+    peer_api = IntroducerAPI(node)
 
     return Service(
         root_path=root_path,
         config=config,
         node=node,
         peer_api=peer_api,
-        node_type=NodeType.TIMELORD,
-        advertised_port=None,
+        node_type=NodeType.INTRODUCER,
+        advertised_port=advertised_port,
         service_name=SERVICE_NAME,
-        connect_peers=get_unresolved_peer_infos(service_config, NodeType.FULL_NODE),
         network_id=network_id,
-        rpc_info=rpc_info,
         connect_to_daemon=connect_to_daemon,
         class_for_type=ApiProtocolRegistry,
     )
@@ -72,7 +71,7 @@ async def async_main(root_path: pathlib.Path) -> int:
     config[SERVICE_NAME] = service_config
     initialize_service_logging(service_name=SERVICE_NAME, config=config, root_path=root_path)
 
-    service = create_timelord_service(root_path, config, DEFAULT_CONSTANTS)
+    service = create_introducer_service(root_path, config)
     async with SignalHandlers.manage() as signal_handlers:
         await service.setup_process_global_state(signal_handlers=signal_handlers)
         await service.run()
