+++ chia/seeder/start_crawler.py
@@ -15,7 +15,7 @@ from chia.protocols.outbound_message import NodeType
 from chia.seeder.crawler import Crawler
 from chia.seeder.crawler_api import CrawlerAPI
 from chia.seeder.crawler_rpc_api import CrawlerRpcApi
-from chia.server.aliases import CrawlerService
+from chia.seeder.crawler_service import CrawlerService
 from chia.server.signal_handlers import SignalHandlers
 from chia.server.start_service import RpcInfo, Service, async_run
 from chia.util.chia_logging import initialize_service_logging
