+++ pyproject.toml
@@ -20,17 +20,17 @@ version = "0.0.0"
 [project.scripts]
 chia = "chia.cmds.chia:main"
 chia_daemon = "chia.daemon.server:main"
-chia_wallet = "chia.server.start_wallet:main"
-chia_full_node = "chia.server.start_full_node:main"
-chia_harvester = "chia.server.start_harvester:main"
-chia_farmer = "chia.server.start_farmer:main"
-chia_introducer = "chia.server.start_introducer:main"
+chia_wallet = "chia.wallet.start_wallet:main"
+chia_full_node = "chia.full_node.start_full_node:main"
+chia_harvester = "chia.harvester.start_harvester:main"
+chia_farmer = "chia.farmer.start_farmer:main"
+chia_introducer = "chia.introducer.start_introducer:main"
 chia_crawler = "chia.seeder.start_crawler:main"
 chia_seeder = "chia.seeder.dns_server:main"
-chia_timelord = "chia.server.start_timelord:main"
+chia_timelord = "chia.timelord.start_timelord:main"
 chia_timelord_launcher = "chia.timelord.timelord_launcher:main"
 chia_full_node_simulator = "chia.simulator.start_simulator:main"
-chia_data_layer = "chia.server.start_data_layer:main"
+chia_data_layer = "chia.data_layer.start_data_layer:main"
 chia_data_layer_http = "chia.data_layer.data_layer_server:main"
 chia_data_layer_s3_plugin = "chia.data_layer.s3_plugin_service:run_server"
