+++ tach.toml
@@ -32,21 +32,8 @@ depends_on = [
     "chia.types",
     "chia.util",
     { path = "chia.ssl", deprecated = false },
-    # these ones can be removed by getting rid of `chia.server.aliases`
-    # and/or the `start_*.py` scripts, which can go in their respective modules
     { path = "chia.rpc", deprecated = false },
-    { path = "chia.full_node", deprecated = false },
-    { path = "chia.data_layer", deprecated = false },
-    { path = "chia.farmer", deprecated = false },
-    { path = "chia.simulator", deprecated = false },
     { path = "chia.daemon", deprecated = false },
-    { path = "chia.harvester", deprecated = false },
-    { path = "chia.timelord", deprecated = false },
-    { path = "chia.introducer", deprecated = false },
-    { path = "chia.consensus", deprecated = false },
-    { path = "chia.wallet", deprecated = false },
-    { path = "chia.apis", deprecated = false },
-    { path = "chia.seeder", deprecated = false },
 ]
 
 [[modules]]
@@ -57,13 +44,16 @@ depends_on = [
     { path = "chia.server", deprecated = false },
     { path = "chia.plotting", deprecated = false },
     { path = "chia.full_node", deprecated = false },
+    { path = "chia.introducer", deprecated = false },
     { path = "chia.wallet", deprecated = false },
+    { path = "chia.farmer", deprecated = false },
     { path = "chia.seeder", deprecated = false },
     { path = "chia.consensus", deprecated = false },
     { path = "chia.apis", deprecated = false },
     { path = "chia.cmds", deprecated = false },
     { path = "chia.protocols", deprecated = false },
     { path = "chia.timelord", deprecated = false },
+    { path = "chia.harvester", deprecated = false },
     { path = "chia.daemon", deprecated = false },
     { path = "chia.ssl", deprecated = false },
     { path = "chia.rpc", deprecated = false },
@@ -72,6 +62,7 @@ depends_on = [
 [[modules]]
 path = "chia.full_node"
 depends_on = [
+    "chia.apis",
     "chia.types",
     "chia.util",
     "chia.consensus",
@@ -83,6 +74,8 @@ depends_on = [
 [[modules]]
 path = "chia.data_layer"
 depends_on = [
+    "chia.apis",
+    "chia.ssl",
     "chia.types",
     "chia.util",
     "chia.protocols",
@@ -108,6 +101,8 @@ depends_on = [
 [[modules]]
 path = "chia.introducer"
 depends_on = [
+    "chia.apis",
+    "chia.full_node",
     "chia.types",
     "chia.util",
     { path = "chia.rpc", deprecated = false },
@@ -128,6 +123,7 @@ depends_on = [
 [[modules]]
 path = "chia.farmer"
 depends_on = [
+    "chia.apis",
     "chia.types",
     "chia.util",
     { path = "chia.protocols", deprecated = false },
@@ -156,6 +152,7 @@ depends_on = ["chia.util"]
 [[modules]]
 path = "chia.timelord"
 depends_on = [
+    "chia.apis",
     "chia.types",
     "chia.util",
     { path = "chia.server", deprecated = false },
@@ -183,6 +180,7 @@ depends_on = [
 [[modules]]
 path = "chia.harvester"
 depends_on = [
+    "chia.apis",
     "chia.types",
     "chia.util",
     { path = "chia.protocols", deprecated = false },
@@ -230,6 +228,8 @@ depends_on = [
 [[modules]]
 path = "chia.wallet"
 depends_on = [
+    "chia.apis",
+    "chia.simulator",
     "chia.types",
     "chia.util",
     { path = "chia.server", deprecated = false },
